backend/requirements.txt
yt-dlp
opencv-python
ultralytics
Pillow
pyyaml
tqdm
python-dotenv
gradio

# FastAPI backend dependencies
fastapi
uvicorn[standard]
websockets
python-multipart
aiofiles

